use `xml2` package instead of `XML` package for `read_epp_data()`
The `XML` package is orphaned and throws a deprecated warning in R 3.4
update. Switch to using `xml2` for the function `read_epp_data()`. The
re-implementation also appears to be somewhat faster (~2x).

There are no feature or data structure changes in this commit, only
under the hood changes. Other functions that use `XML` should be
updated in future commits.